test: fix bootstrap probe autoload
