+++ CHANGELOG.md
@@ -37,6 +37,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- moved the bootstrap environment file-name override test probe out of `tests/Unit/TestCaseBootstrapEnvironmentFileTest.php` file scope into a PSR-4 compliant `Tests\Support\...` helper so `composer install` / optimized autoload generation no longer emits a `Class TestCaseBootstrapEnvironmentFileNameOverrideProbe ... does not comply with psr-4 autoloading standard` warning during development and CI bootstrap
 - isolated local API test bootstrap from deployment-oriented `.env` drift by generating a dedicated test-only env file for `php artisan test`, ignoring repository `BOOTSTRAP_*` deployment flags during suite bootstrap-default assertions, and restricting intentional local env passthrough to PostgreSQL connection keys (`DB_HOST`, `DB_PORT`, `DB_USERNAME`, `DB_PASSWORD`) while providing a deterministic test `APP_KEY` plus updated local test-environment documentation (refs `api#1148`)
 - adopted a shared `App\Support\ApiTimestamp` serializer across API resources, controllers, and service-built response payloads so canonical response timestamps now consistently normalize to UTC whole-second strings with a trailing `Z`; notification-installation timestamps now use the same contract-approved serializer and focused Pest coverage locks the format in place (refs `api#1138`)
 - fixed public bootstrap `api_base_url` derivation so `APP_URL` values with unrelated path prefixes now fail closed instead of publishing an invented nested API path, while `APP_URL` values already ending in `/v1` continue to resolve to a single canonical `/v1` base URL
