+++ tests/Unit/TestCaseBootstrapEnvironmentFileTest.php
@@ -6,53 +6,8 @@
 declare(strict_types=1);
 
 use Illuminate\Support\Str;
+use Tests\Support\TestCaseBootstrapEnvironmentFileNameOverrideProbe;
 use Tests\Support\TestCaseBootstrapEnvironmentProbe;
-use Tests\TestCase;
-
-final class TestCaseBootstrapEnvironmentFileNameOverrideProbe extends TestCase
-{
-    private static ?string $probeEnvironmentFileName = null;
-
-    public static function useProbeEnvironmentFileName(string $fileName): void
-    {
-        self::$probeEnvironmentFileName = $fileName;
-    }
-
-    public static function clearProbeEnvironmentFileName(): void
-    {
-        self::$probeEnvironmentFileName = null;
-    }
-
-    public static function createBootstrapEnvironmentStub(): void
-    {
-        parent::prepareBootstrapEnvironment();
-    }
-
-    public static function removeBootstrapEnvironmentStub(): void
-    {
-        parent::cleanupBootstrapEnvironmentFile();
-    }
-
-    public static function resetBootstrapEnvironmentState(): void
-    {
-        parent::resetBootstrapEnvironmentState();
-    }
-
-    public static function bootstrapEnvironmentFilePath(): string
-    {
-        return parent::bootstrapEnvironmentFilePath();
-    }
-
-    public static function createBootstrapApplication(): Illuminate\Foundation\Application
-    {
-        return (new self('createApplication'))->createApplication();
-    }
-
-    protected static function bootstrapEnvironmentFileName(): string
-    {
-        return self::$probeEnvironmentFileName ?? parent::bootstrapEnvironmentFileName();
-    }
-}
 
 afterEach(function (): void {
     TestCaseBootstrapEnvironmentProbe::removeBootstrapEnvironmentStub();
